Replaced gulp-file-include with gulp-nunjucks
