--- combined.json
@@ -1,15 +1,0 @@
-{
-	"title": "This is the title",
-	"description": "This is the description",
-	"en": {
-		"title": "English title",
-		"age": 333
-	},
-	"de": {
-		"title": "German title",
-		"age": 456
-	},
-	"context": {
-		"task": "build"
-	}
-}
