+++ config.json
@@ -6,7 +6,7 @@
 			"styles": "src/styles/**/*.scss",
 			"stylesFolder": "src/styles",
 			"spriteSrc": "src/images/sprite-src",
-			"html": ["src/pages/**/*.html", "!src/includes/**/*.html"],
+			"html": ["src/pages/**/*.html", "!src/pages/includes/**/*.html"],
 			"htmlWatch": ["src/**/*.html"],
 			"images": "src/images/**/*",
 			"static": ["src/**/*", "!src/**/*.html", "!src/scripts/**/*", "!src/styles/**/*", "!src/data/**/*"],
