+++ gulpfile.js
@@ -32,8 +32,7 @@ require('./gulp-tasks/es-lint')(gulp, plugins, config, errorHandler);
 require('./gulp-tasks/image-min')(gulp, plugins, config, errorHandler);
 
 // HTML templating
-require('./gulp-tasks/html-templating-develop')(gulp, plugins, config, errorHandler);
-require('./gulp-tasks/html-templating-build')(gulp, plugins, config, errorHandler);
+require('./gulp-tasks/html-templating')(gulp, plugins, config, errorHandler);
 
 // HTML standards
 require('./gulp-tasks/html-lint')(gulp, plugins, config, errorHandler);
@@ -46,37 +45,6 @@ require('./gulp-tasks/watch')(gulp, plugins, config, errorHandler);
 
 // Complete
 require('./gulp-tasks/complete')(gulp, plugins, config, errorHandler);
-
-gulp.task('merge', function(callback) {
-	plugins.runSequence('clean', 'merge-data', 'compile-data');
-});
-
-var combined = {};
-
-gulp.task('merge-data', function (callback) {
-	gulp.src(config.paths.input.data)
-		.pipe(plugins.mergeJson({
-			endObj: {
-				'context': {
-					'task': 'build'
-				}
-			}
-		}))
-		.pipe(gulp.dest((output) => {
-			combined = JSON.parse(output.contents.toString());
-
-			callback();
-			return '';
-		}));
-});
-
-gulp.task('compile-data', function (callback) {
-	gulp.src(config.paths.input.html)
-		.pipe(plugins.nunjucks.compile(combined))
-		.pipe(gulp.dest(config.paths.output.devRoot));
-
-	callback();
-});
 
 
 gulp.task('develop', function(callback) {
