+++ package.json
@@ -16,7 +16,6 @@
     "gulp-clean": "~0.3.2",
     "gulp-clean-css": "~3.7.0",
     "gulp-eslint": "~4.0.0",
-    "gulp-html-templating": "~0.1.8",
     "gulp-htmllint": "0.0.12",
     "gulp-imagemin": "~3.3.0",
     "gulp-load-plugins": "~1.5.0",
