+++ src/data/gamma.json
@@ -1,9 +1,11 @@
 {
-	"de": {
-		"title": "German title",
-		"age": 456
-	},
-	"en": {
-		"age": 333
+	"translations": {
+		"de": {
+			"title": "German title",
+			"age": 456
+		},
+		"en": {
+			"age": 333
+		}
 	}
 }
