--- src/includes/sample.html
@@ -1,1 +1,0 @@
-<h2>@@title</h2>
