+++ src/pages/404.html
@@ -7,15 +7,21 @@
 		<meta name="viewport" content="width=device-width, initial-scale=1">
 		<link rel="apple-touch-icon" href="icon.png">
 
-		@@if (task === 'build') {<link rel="stylesheet" type="text/css" href="styles/main.min.css" media="screen">}
-		@@if (task !== 'build') {<link rel="stylesheet" type="text/css" href="styles/main.css" media="screen">}
+		{% if content.task === 'build' %}
+			<link rel="stylesheet" type="text/css" href="styles/main.min.css" media="screen">
+		{% else %}
+			<link rel="stylesheet" type="text/css" href="styles/main.css" media="screen">
+		{% endif %}
 	</head>
 	<body>
 		<h1>Page not found</h1>
 
-		@@include("../includes/footer.html")
+		{% include "./includes/footer.html" %}
 
-		@@if (task === 'build') {<script src="scripts/main.min.js"></script>}
-		@@if (task !== 'build') {<script src="scripts/main.js"></script>}
+		{% if content.task === 'build' %}
+			<script src="scripts/main.min.js"></script>
+		{% else %}
+			<script src="scripts/main.js"></script>
+		{% endif %}
 	</body>
 </html>
