+++ src/pages/includes/sample.html
@@ -1,0 +1,7 @@
+
+{% macro translateTitle(content) %}
+	<h2>{{ content.title }}</h2>
+{% endmacro %}
+
+{{ translateTitle(translations.en) }}
+{{ translateTitle(translations.de) }}
