+++ src/pages/index.html
@@ -7,26 +7,31 @@
 		<meta name="viewport" content="width=device-width, initial-scale=1">
 		<link rel="apple-touch-icon" href="icon.png">
 
-		@@if (task === 'build') {<link rel="stylesheet" type="text/css" href="styles/main.min.css" media="screen">}
-		@@if (task !== 'build') {<link rel="stylesheet" type="text/css" href="styles/main.css" media="screen">}
+		{% if content.task === 'build' %}
+			<link rel="stylesheet" type="text/css" href="styles/main.min.css" media="screen">
+		{% else %}
+			<link rel="stylesheet" type="text/css" href="styles/main.css" media="screen">
+		{% endif %}
 	</head>
 	<body>
+		<div class="container">
+			<div class="row">
+				<div class="col-sm-12">
+					<h1 class="h1">{{ title }}</h1>
 
-Hello {{ title }}
-
-	<div class="container">
-		<div class="row">
-			<div class="col-sm-12">
-				<h1 class="h1">Hello worldddd</h1> @@include("../includes/sample.html", { "title": "This is a title" })
+					{% include "./includes/sample.html" %}
+				</div>
 			</div>
 		</div>
-	</div>
 
-	@@include("../includes/grid-examples.html")
+		{% include "./includes/grid-examples.html" %}
 
-		@@include("../includes/footer.html")
+		{% include "./includes/footer.html" %}
 
-		@@if (task === 'build') {<script src="scripts/main.min.js"></script>}
-		@@if (task !== 'build') {<script src="scripts/main.js"></script>}
+		{% if content.task === 'build' %}
+			<script src="scripts/main.min.js"></script>
+		{% else %}
+			<script src="scripts/main.js"></script>
+		{% endif %}
 	</body>
 </html>
